Update to 2018 / Pad to byte / Documentation update
The main purpose of the update was to add the functionality of pad_to_byte. Without this, mainting the state of bits written would have to be done at a program level. The necessary value 'shift' was already stored. Functionality should be added at the bit-writer level.

Update code standards to current rust 1.33/2018. This mainly consisted of removing try! macro usage.

Updated documentation to reflect above changes. Fixed errors in documentation where bw was used instead of br so the code would not compile if tested. Made br/bw usage consistent.